Fix King Prefix error
Holy shit I added a space. I implied that there was supposed to be a space in the code.

This may be a bit of couplin and maybe should be left to the prefix.

I'll sort it out another day. And never do it anyways.

Gonna have this commit haunt me one day.

Yee yee.

I am still alive and it is August in 2020.

I do hope the Russian Vaccine works.

Ramen.